7z-sys-provided: let user override paths
let the user override assumed paths via environment.
EXTRACTCODE_7Z_PATH is used to determine the path of the 7z
executable.
If not specified it falls back to previously used probe of
distro data.

With this patch it possible to run scancode is buildsystems like YOCTO,
which heaviily rely on overriding paths via environment to pick the
correct implementation from the build workspace

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>