libmagic-sys-provided: let user override paths
let the user override assumed paths via environment.
TYPECODE_LIBMAGIC_PATH environment variable can be used to explicitly
set the path to libmagic.so. If not specified it falls back to
previously used calculation from distro data.
Location of the magic database can be specified via
TYPECODE_LIBMAGIC_DB_PATH environment variable.
If not found it falls back to the previous bevahior.

With this patch it possible to run scancode is buildsystems like YOCTO,
which heaviily rely on overriding paths via environment to pick the
correct implementation from the build workspace

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>