Make SevenzipPaths.get_locations() more robust
Current Fedora Rawhide (which is to become version 43 eventually) has
the proper 7zip package, which in turn doesn’t have the tool in the
place expected for prior versions (with p7zip-plugins).

To make the method more robust, try all known directories and binary
names on Linux, but make the order in which they are checked dependent
on the type of Linux distribution encountered.

Signed-off-by: Nils Philippsen <nils@redhat.com>